ECS refactor
added `specs` crate as dependency
basically... as I was working on level 2, I ran into a wall. hard.
ruby warrior is coded in such a way that there is bidirectional
referencing (with mutations). rust REALLY does not want me to do
that -- I was unable to compile (when I had almost finished the
level) because I could either hold an immutable and mutable
reference to the same thing, or two mutable references to the
same thing. I was trying to set up `Level` to manage the different
units as "components", but it proved very difficult.

I knew about `specs`, and I had a feeling that it would work really
well in this game. I did a POC to try it out, and it went great!
The POC even has a monster and there is attacking and collision
detection. So those things will be coming very soon!

First, I need to figure out how to initialize `specs` with the
details for each level. That way I can start it up for level 1,
and then set it up differently for level 2.