level 2 complete
in rust-warrior, level 2 introduces the first enemy unit - the Sludge.
this is where ECS comes in. we want a way to manage all this changing
state:

- warrior position, health
- sludge position, health

a requirement is that we're able to read the warrior and sludge's
positions to determine if they're in range of each others' attacks.

another requirement is that the sludge is an optional unit. i.e. he
does not exist on level 1, so the game engine needs to allow that
flexibility.

a slightly tricky part is presenting the player with a
straightforward interface for controlling the warrior. at first, I
tried wiring the game engine into the `Warrior` object. a lightbulb
moment occurred, though. now there is a simple `Action` enum to
inform the engine of what the player decided to do for each turn.

within the engine, I started out with very fine-grained components:
`Position`, `HP`. but I quickly decided to combine these. next, I
tried having a `Warrior` and `Sludge` component. I liked how you
could query for them specifically. unfortunately, specs makes it
difficult to query entirely separate components at the same time.
so the final design was to have a `Unit` component and a `UnitType`
enum to tell them apart. this will probably work nicely once there
are other types of units.

right now the `Player` system spends way too much energy trying to
mess with the Sludge (likewise the `Sludge` system is what changes
the warrior's health after an attack). so I'll need to come up with
a better approach for detecting collision when walking and dealing
damage when attacking. there could maybe be `Attack` or `Movement`
systems for this, but I'm not sure how the position would get
synchronized, etc. I'll have to fiddle with that.