simplify player name lookup
originally I considered this approach, and I should have just done it.
it seemed like it might be easier to just stuff the name into the
UnitType enum, but now I am considering exposing that to the player
(`check` returning a `Tile` which can be a `Unit(UnitType)` variant).
also, generally speaking, everything was just getting way more messy
by taking away the simplicity of the `UnitType` enum (losing the `Copy`
trait, losing the ability to do `==` on the `Warrior` variant, etc)