fix broken iterators in specs systems
from time to time, I've noticed some weird behavior with the
`join` iterators in these Systems, but until now I've been able
to work around it.

the problem - I was trying to iterate multiple times over the
Iterator returned by the specs `join` method, to find certain
components that exist in the world. there was a conveniently
hidden bug due to the warrior always being the first one
returned by these iterators. since the warrior was always added
to the world first, and then also extracted first in these systems
(using `find`, which short-circuits), the iterator only had to
advance one position and then behaved fine when it was used again
to extract other components.

on level 6, though, the warrior is no longer the first component
added to the world. so by traversing the iterator to reach the
warrior, we have lost any component added to the world prior to
the warrior - in this case, the captive to his left.

the simple solution is to do a single loop on the `join`
iterator and save references to all of the components encountered.
this also has the added benefit of being (probably) quite a bit
more readable too.