tuning for level 4
basically just stretched it by one tile

it wasn't technically broken, but I have a special use case... while developing
new levels I tend to use a single Player definition to replay all of the levels
(to ensure that I haven't managed to break a previous level).

SPOILER: in order to complete level 6, I instructed the warrior to back up
against the west wall to rest if there were archers shooting and the warrior's
HP was low. on level 4 if you attempted this, you could be backed up against the
wall and still get hit by the archer. I think I increased the archer range when
implementing level 6, and that may have also had an influence on this.